Fix the cell standby mobile radio active issue
How to reproduce the bug:
-Enable mobile data
-Use some mobile data (radio active timer starts in battery stats)
-Disable mobile data (radio active timer keeps running in battery stats)

After some investigation on the issue I found that the BatteryStats
service was not receiving the radio power down notification. The
investigation lead me to the NetworkManagementService.java, where
I found that some code was discarding the radio power change
notifications after the first radio power on.

The problem is that in NetworkManagementServer.java the
notifyInterfaceClassActivity method is discarding the notifications
of radio power change to the BatteryStats (via noteMobileRadioPowerState)
if the fromRadio parameter is false after it's being true some time before.

Even after a detailed analysis, it was unclear what is the real role of
the fromRadio parameter. Changing it to true in the calling methods
solved the issue in the testing unit (Sony Xperia ZL). I started a thread
in XDA to mobilize some volunteers to help test this fix.

http://forum.xda-developers.com/xperia-zl/general/cm12-1-cell-standby-mobile-radio-active-t3188147

Change that introduced the code with the potential to trigger the bug:
 -Introduction of "boolean fromRadio" in
  NetworkManagementService.java / notifyInterfaceClassActivity()
 https://android.googlesource.com/platform/frameworks/base/+/2ffa11e4b71c545e34533ef827bdc1a07fbe8246

Change that triggered the bug:
 -Uncommenting the code that calls notifyInterfaceClassActivity
  with fromRadio=true
 https://android.googlesource.com/platform/frameworks/base/+/2d4b4ed681efcfb64c4d2b4d2a2d236aae1dc7ee

Change-Id: Icec9db692b0781dfb869858796ca8bb448db3f62
Signed-off-by: Chet Kener <Cl3Kener@gmail.com>