Make sure NaturalSurfaceLayout is consistent with SurfaceControl
SurfaceControl and NaturalSurfaceLayout will be created in screen-off
mode. Screen-on animation also creates NaturalSurfaceLayout. And
NaturalSurfaceLayout uses the reference of SurfaceControl.

SurfaceControl points to NULL when screen-on process calls destory
function, which will not be called in screen-off process. It makes
NaturalSurfaceLayout's member points to the SurfaceControl with NULL.
Crash occurs by accessing the destroyed SurfaceControl.

This CL makes the creation of SurfaceControl and NaturalSurfaceLayout
to one-to-one.

Change-Id: Ic47ea73dea67d2aba4e3e0766d1353198d7dff1d
Signed-off-by: Chet Kener <Cl3Kener@gmail.com>